Final fix of issue #21, for all parts of speech
In addition to adding clarification comments, fixed the problem created
by fixing issue #21 (adding a lexical pointer field):
-converter_noun.py - added a default lexical pointer for preexisting
nouns (its value, -1, is defined in converter_row.py, as it is identical
for all converters). Added a comparison with preexisting keys, so as not
to add a new key that is identical to a preexisting one, except its
lexical pointer.
-converter_verb.py - same changes as for nouns.
-converter_adjective.py - same changes as for nouns.
-converter_adverb.py - similar changes as for nouns, only here the data
structure is a set and not a dictionary.
As I only had examples of issue #21 for nouns, that's the only case that
I could make sure was fixed. But it should have been, as it's the same
solution. As for the fix of the resulting issue - it has been
successfully QAed for all parts of speech.