Many changes and fixes to converter_verb.py
1) "Special" pre-existing verbs are now also stored in a separate
dictionary, so that they can be queried more efficiently (when
preventing multiple entries for the same STEM & PRED combinations). This
is a subpart of bug #24 (most of this bug  was solved by adding the
lexical pointer, see commit called "final fix of issue #21").
2) Made the treatment of the pre-existing unaccusative types more
efficient.
3) Fixed typo in "imperitive"
4) Fixed bug #19 - verbs that should have multiple entries, now get
them, and then fixed the resulting bug #23 (multiple control entries for
verbs).