Fix issue #27 - get rid of illegal punctuation marks in STEM/PRED values
This is done via the new "replacePunct" method in
"converter_helpers.py".
All four converters (adjective, adverb, noun, and verb) call it, and
then create a key in their dictionary with the "clean" STEM and PRED
values.  " is replaced with 2, ' is replaced with 1, and . is replaced
with 0.
Therefore, there is no need to separately print out the two types of
output (i.e., the punctuation lexica are now defunct, as seen in
CSVToTDLLexicalType.py).
Note - the adverb file only appears to have undergone many changes, most
of these are indentation corrections.