Fixing issues #31, #32, #33 and adding debug helper
-In both files: added an option where there's a dictionary, based on
PMI_dictionary, with the lexical pointer as the key, and possible
complements as the value (in a list). This gives a way to check if some
combinations of complements appear in the input at all (so as to
understand if it's ok that they aren't produced). So this "solved"
issues #34 :-) At this point, this is commented out but can be used if
needed.
-In converter_verb: added specific cases for needed combinations, to
solve issues #31, #32, #33 - basically, when entries weren't merged when
they needed to be.